requirements.txt
Flask==3.0.3
flask-cors==4.0.1
openai==1.14.0
python-dotenv==1.0.1
httpx==0.27.0